+++ service-user/src/main/java/by/salary/serviceuser/service/UserAgreementsService.java
@@ -1,11 +1,13 @@
 package by.salary.serviceuser.service;
 
 import by.salary.serviceuser.entities.UserAgreement;
+import by.salary.serviceuser.exceptions.UserNotFoundException;
 import by.salary.serviceuser.model.UserAgreementRequestDTO;
 import by.salary.serviceuser.model.UserAgreementResponseDTO;
 import by.salary.serviceuser.repository.UserAgreementsRepository;
 import by.salary.serviceuser.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.math.BigInteger;
@@ -26,20 +28,31 @@ public class UserAgreementsService {
 
 
     public List<UserAgreementResponseDTO> getAllUserAgreements(BigInteger userId) {
+        if(!userRepository.existsById(userId)){
+            throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
+        }
+
         List<UserAgreementResponseDTO> userAgreements = new ArrayList<>();
         userAgreementsRepository.findAll().forEach(userAgreement -> {
             if (userAgreement.getUser().getId().equals(userId)) {
                 userAgreements.add(new UserAgreementResponseDTO(userAgreement));
             }
         });
+
         return userAgreements;
     }
 
     public void deleteUserAgreement(BigInteger agreementId) {
+        if(!userAgreementsRepository.existsById(agreementId)){
+            throw new UserNotFoundException("User agreement with id " + agreementId + " not found", HttpStatus.NOT_FOUND);
+        }
         userAgreementsRepository.deleteById(agreementId);
     }
 
     public void deleteUserAgreements(BigInteger userId) {
+        if(!userRepository.existsById(userId)){
+            throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
+        }
         userAgreementsRepository.findAll().forEach(
                 userAgreement -> {
                     if (userAgreement.getUser().getId().equals(userId)) {
@@ -50,6 +63,9 @@ public class UserAgreementsService {
     }
 
     public UserAgreementResponseDTO createUserAgreement(UserAgreementRequestDTO userAgreementRequestDTO, BigInteger userId) {
+        if(!userRepository.existsById(userId)){
+            throw new UserNotFoundException("User with id " + userId + " not found", HttpStatus.NOT_FOUND);
+        }
         UserAgreement userAgreement = new UserAgreement(userAgreementRequestDTO, userRepository.findById(userId).get());
         return new UserAgreementResponseDTO(userAgreementsRepository.save(userAgreement));
     }
